Filter Wikitionary found IPA based on current language
We load the languages names corresponding to the current language code (will be set globally in the plugin in the future).
These names are already in a regular expression format as LangName 1|LangName 2.
We use the language name to select the only the IPA previously found, which correspond to the current language.