Add callback to radio button selection
When a radio button is clicked to select a pronunciation the method 'setPronunciation' is called and the audio is saved in the note. The javascript function assigned to onclick and provided by _myscript is bridged to setPronunciation by a so called 'bridge' provided in anki.editor, we'll need to extend the bridge to make it work.