Add article to note automatically based on list of words
For danish:
We compare the word with a list of word loaded from a file, if the word is in the list we know that the word is of gender 'neuter' and we  add the article et.

This should be extended in a better way. I think it doesn't even work know.