Add button for fast note cloning
When we use the add on note id we can create duplicate notes for the same word, this can be useful to distinguish between different meanings of the same word (some words are nouns and verbs).

It doesn't work for now
ToDO: What if the user doesn't have the add on NoteId installed?