Add link handling within IFrame
Open the link within the correct IFrame (when more than one IFrame is present). If the link is not coming from an IFrame, try to open it in the system browser.

FIXME: Does it make sense to open in the system browser?
The idea was to open in Chrome to use the translation feature on the page