Set identifiers of ipa entries in the form ipa[0-9]
Before we were passing the ipa text but because the anki bridge tries to print the string there were an exception because of unicode.

Now we pass an identifier in the form ipa[0-9] and ipac[0-9] for current ipas.

Moreover we manage the case when more than a entry is selected.