Extract noise from track
If the audio track is long enough (3 times the specified noise length) we cut a bit of it from the end (usually there is more noisy silence at the end of the track) to have a sample of pure noise.

noiseSampleLength should be little enough to avoid of including speech in the cut portion but big enough to have enough noise samples.