Fix missing images when link to full resolution image is broken
To avoid this problem a check on the downloaded image is performed to see if it's valid, otherwise the thumbnail is used.

This should not be a problem if the thumbnail resolution is high enough, but we still warn the user with a message box.

Fix issue in Trello card #49