Add buttons with new ani filter
The buttons are now written in the HTML so we need to add the button as html buttons. Also the method toggleVocabularyDeckBuilder must be added to the _link dictionary of anki so that it can be called properly when the button is pressed.

Note: This assumes that the pull request for the anki filter is accepted, otherwise this code will need to be adapted